Refactor conda.common.pkg_formats.python
