+++ conda/common/pkg_formats/__init__.py
@@ -1,2 +1,6 @@
 # Copyright (C) 2012 Anaconda, Inc
 # SPDX-License-Identifier: BSD-3-Clause
+
+from ...deprecations import deprecated
+
+deprecated.module("25.9", "26.3", addendum="Use 'prefix_data_loaders' hook instead")
